Include table xref code in YAML
We previously only had a boolean field to indicate whether a column was
reliant on a foreign key present in file 1 (the table file). This was
retaining the logic present in the SCC’s file, which just uses a “+” to
indicate the same thing. The difficulty is that there is no mechanism
to determine *which* foreign key was to be used. Previously we’d tried
renaming the column to use the same identifier that’s found in table 1,
but that’s troublesome—it leads to less-descriptive column names,
swapping out less precise terminology.

Now we’re improving on the presence of the boolean field (which wasn’t
even really boolean before—instead of “false,” the field just wouldn’t
be included) by including the table code, so that there’s no question
as to which foreign key should be used.